requirements.txt
absl-py==0.15.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.10.8
cycler==0.11.0
fonttools==4.29.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
importlib-metadata==4.8.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.4
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
pip==21.2.2
protobuf==3.19.1
pyparsing==3.0.7
python-dateutil==2.8.2
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing-extensions==3.10.0.2
webencodings==0.5.1
Werkzeug==0.16.1
wheel==0.37.1
wrapt==1.13.3
zipp==3.7.0
requests==2.28.2
flask==1.1.4
markupsafe==2.0.1